Partially fixed microbe editor visuals
Fixed some editor GUI stuff

+++ src/general/base_stage/HexEditorComponentBase.cs
@@ -413,7 +413,7 @@ public partial class HexEditorComponentBase<TEditor, TCombinedAction, TAction, T
     }
 
     [RunOnKey("e_pan_mouse", CallbackRequiresElapsedTime = false, OnlyUnhandled = false)]
-    public bool PanCameraWithMouse(float dummy)
+    public bool PanCameraWithMouse(double dummy)
     {
         _ = dummy;
